Added support for item value formatting Converted itemvalue to double instead of int Fixed item arrangement bug caused by hashmap Fixed off pixel due to rounding off of int Updated example

+++ lib/src/main/java/com/chuck/multicolorbar/MulticolorBarItem.java
@@ -6,7 +6,7 @@ package com.chuck.multicolorbar;
 
 public interface MulticolorBarItem {
 
-    int getItemValue();
+    double getItemValue();
     String getItemName();
     String getColorHex();
     String getUnit();
